move test to heal_test.go file
